Rename defillama stablecoins pull to make room for new version based on GCS
