+++ src/op_analytics/cli/subcommands/pulls/app.py
@@ -9,7 +9,7 @@ from .defillama.historical_chain_tvl import (
     pull_historical_chain_tvl as dfl_pull_historical_chain_tvl,
 )
 from .defillama.protocols import pull_protocol_tvl as dfl_pull_protocol_tvl
-from .defillama.stablecoins import pull_stablecoins as dfl_pull_stablecoins
+from .defillama.stablecoins_bigquery import pull_stablecoins as dfl_pull_stablecoins
 from .github_analytics import pull as github_analytics_pull
 from .l2beat import pull as l2beat_pull
