+++ src/op_analytics/cli/subcommands/pulls/defillama/stablecoins_bigquery.py
