+++ tests/op_analytics/cli/pulls/defillama/test_defillama_stablecoins.py
@@ -4,7 +4,7 @@ import pytest
 from op_analytics.coreutils.testutils.inputdata import InputTestData
 
 
-from op_analytics.cli.subcommands.pulls.defillama import stablecoins
+from op_analytics.cli.subcommands.pulls.defillama import stablecoins_bigquery
 
 TESTDATA = InputTestData.at(__file__)
 
@@ -90,7 +90,7 @@ another_sample_breakdown_data = {
 
 def test_process_breakdown_stables():
     # Test with valid data
-    metadata, balances = stablecoins.single_stablecoin_balances(sample_breakdown_data)
+    metadata, balances = stablecoins_bigquery.single_stablecoin_balances(sample_breakdown_data)
     assert len(balances) == 2  # Two data points within default 30 days
     assert metadata["id"] == "sample-stablecoin"
     assert metadata["name"] == "Sample Stablecoin"
@@ -105,7 +105,7 @@ def test_process_breakdown_stables():
     incomplete_data = sample_breakdown_data.copy()
     del incomplete_data["pegType"]
     with pytest.raises(KeyError) as excinfo:
-        stablecoins.single_stablecoin_balances(incomplete_data)
+        stablecoins_bigquery.single_stablecoin_balances(incomplete_data)
     assert excinfo.value.args == ("pegType",)
 
 
@@ -124,7 +124,7 @@ def test_pull_stables_single_stablecoin(
     ]
 
     # Call the function under test
-    result = stablecoins.pull_stablecoins(symbols=["SSC"])
+    result = stablecoins_bigquery.pull_stablecoins(symbols=["SSC"])
 
     # Assertions
     assert len(result.metadata_df) == 1  # Only 'sample-stablecoin'
@@ -163,7 +163,7 @@ def test_pull_stables_multiple_stablecoins(
     ]
 
     # Call the function under test without specifying stablecoin_ids (process all)
-    result = stablecoins.pull_stablecoins()
+    result = stablecoins_bigquery.pull_stablecoins()
 
     # Assertions
     assert len(result.metadata_df) == 2  # Both stablecoins
@@ -196,7 +196,7 @@ def test_pull_stables_no_valid_ids():
     ) as mock_get_data:
         mock_get_data.return_value = sample_summary
         with pytest.raises(ValueError) as excinfo:
-            stablecoins.pull_stablecoins(symbols=["NOEXIST"])
+            stablecoins_bigquery.pull_stablecoins(symbols=["NOEXIST"])
         assert "No valid stablecoin IDs provided." in str(excinfo.value)
 
 
@@ -207,7 +207,7 @@ def test_pull_stables_missing_pegged_assets():
     ) as mock_get_data:
         mock_get_data.return_value = {}
         with pytest.raises(KeyError) as excinfo:
-            stablecoins.pull_stablecoins()
+            stablecoins_bigquery.pull_stablecoins()
         assert excinfo.value.args == ("peggedAssets",)
 
 
@@ -215,7 +215,7 @@ def test_process_breakdown_stables_empty_balances():
     # Test process_breakdown_stables with empty 'chainBalances' (should return empty DataFrame)
     data_with_empty_balances = sample_breakdown_data.copy()
     data_with_empty_balances["chainBalances"] = {}
-    metadata, balances = stablecoins.single_stablecoin_balances(data_with_empty_balances)
+    metadata, balances = stablecoins_bigquery.single_stablecoin_balances(data_with_empty_balances)
     assert not balances
     assert metadata["id"] == "sample-stablecoin"
 
@@ -225,7 +225,7 @@ def test_process_breakdown_stables_missing_mandatory_metadata():
     incomplete_data = sample_breakdown_data.copy()
     del incomplete_data["name"]  # Remove a mandatory field
     with pytest.raises(KeyError) as excinfo:
-        stablecoins.single_stablecoin_balances(incomplete_data)
+        stablecoins_bigquery.single_stablecoin_balances(incomplete_data)
     assert excinfo.value.args == ("name",)
 
 
@@ -233,7 +233,7 @@ def test_process_breakdown_stables_optional_metadata():
     # Test process_breakdown_stables with missing optional metadata fields
     incomplete_data = sample_breakdown_data.copy()
     del incomplete_data["description"]  # Remove an optional field
-    metadata, balances = stablecoins.single_stablecoin_balances(incomplete_data)
+    metadata, balances = stablecoins_bigquery.single_stablecoin_balances(incomplete_data)
     assert "description" in metadata
     assert metadata["description"] is None  # Should be None if missing
 
@@ -258,7 +258,7 @@ def test_pull_stables_empty_metadata_df():
             },
         ]
         with pytest.raises(ValueError) as excinfo:
-            stablecoins.pull_stablecoins(symbols=["SSC"])
+            stablecoins_bigquery.pull_stablecoins(symbols=["SSC"])
         assert excinfo.value.args == ("No balances for stablecoin=Sample Stablecoin",)
 
 
@@ -281,5 +281,5 @@ def test_pull_stables_empty_breakdown_df():
             },
         ]
         with pytest.raises(ValueError) as excinfo:
-            stablecoins.pull_stablecoins(symbols=["SSC"])
+            stablecoins_bigquery.pull_stablecoins(symbols=["SSC"])
         assert excinfo.value.args == ("No balances for stablecoin=Sample Stablecoin",)
